\name{vgcontainer}
\alias{vgcontainer}
\title{Grouping graphical objects}
\usage{
vgcontainer(..., type = "g")
}
\arguments{
\item{\dots}{none or some "vg*" objects.}

\item{type}{string for \acronym{SVG} container name, one of "g" (group), "svg",
"symbol", "marker".}
}
\description{
Create a container for graphical objects.
}
\details{
Grouping allows for applying common transformations or styles to many objects at once.
}
